+++ src/main/java/org/bytedeco/javacpp/tools/ParserException.java
@@ -29,4 +29,5 @@ package org.bytedeco.javacpp.tools;
 public class ParserException extends Exception {
     public ParserException(String message) { super(message); }
     public ParserException(String message, Throwable cause) { super(message, cause); }
+    public ParserException(Throwable cause) { super(cause); }
 }